COMMERCE-2820 - move to error package
